Express coordinates relative to local CRS
Major refactor of the logic to enable storage of coordinates with
respect to a local CRS. The tree now holds a reference to the transform
objects to be used to convert to/from the epsg 4978 and the local crs
system. The tileset leverages the transform properties to transform
coordinates transparently.