Fix race condition in combined LAS reader
Combined LAS reader used to be called from a single goroutine. From
V2.0.0 final this is no longer the case, however the GetNext() function
was not made thread safe. This change fixes the race condition via CAS
operations and also solves an infinite deadlock due to a lack of lock
release on error